final_project/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
colorama==0.4.6
coverage==7.10.7
exceptiongroup==1.3.1
frozenlist==1.8.0
gigachat==0.1.43
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
magic-filter==1.0.12
multidict==6.7.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dotenv==1.2.1
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
